#unit-tests mock example, fix type

+++ src/Utils/Formatter.php
@@ -33,7 +33,7 @@ class Formatter
      * @param $lost
      * @return int
      */
-    protected function roundedPercentage($won, $lost): int
+    protected function roundedPercentage($won, $lost): float
     {
         return Math::roundedPercentage($won, $lost);
     }
